requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-dotenv
pydantic

numpy
scikit-learn
scipy
rank-bm25
rapidfuzz

pypdfium2
pymupdf
pytesseract

faiss-cpu
chromadb

google-generativeai
huggingface_hub
transformers

sentence-transformers